Like entity is added in the backend logic and based on that post layers are modified

+++ blogAppBackend/src/main/java/com/blogApplication/blogApp/entities/Post.java
@@ -42,7 +42,7 @@ public class Post {
     private String imageUrl;
 
     @Column
-    private Long likeCount = 0L;
+    private Integer likeCount = 0;
 
 
     @CreationTimestamp
